server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
google-generativeai==0.3.1
json-repair==0.7.0
gspread==5.12.0
google-auth==2.25.2
python-multipart==0.0.6
